Refactor helper script to not use exec command

+++ modules/helper/manifests/script.pp
@@ -1,34 +1,20 @@
-define helper::script ($content, $unless = false) {
+define helper::script ($content, $unless) {
 
 	$scriptName = md5($title)
 	$scriptDirname = "/tmp/${scriptName}"
-	$scriptFilename = "${scriptDirname}/${scriptName}"
+	$contentQuoted = shellquote($content)
 
-	unless $unless {
+	exec {"exec ${title}":
+		command => "mkdir -p ${scriptDirname} && cd ${scriptDirname} && echo ${contentQuoted} > ${scriptName} && chmod +x ${$scriptName} && ./${$scriptName}",
+		unless => $unless,
+		path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin'],
+		logoutput => on_failure,
+	}
 
-		file { $scriptDirname:
-			ensure => directory,
-			group => 0, owner => 0, mode => 644,
-		}
-
-		file { $scriptFilename:
-			content => $content,
-			ensure => present,
-			group => 0, owner => 0, mode => 755,
-			require => File[$scriptDirname],
-		}
-
-		exec {"exec ${title}":
-			command => "/bin/bash -ec '${scriptFilename}'",
-			cwd => $scriptDirname,
-			path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin'],
-			require => File[$scriptFilename],
-		}
-
-		exec {"cleanup ${title}":
-			command => "rm -rf ${scriptDirname}",
-			path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin'],
-			require => Exec["exec ${title}"],
-		}
+	exec {"cleanup ${title}":
+		command => "rm -rf ${scriptDirname}",
+		onlyif => "test -d ${scriptDirname}",
+		path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin'],
+		require => Exec["exec ${title}"],
 	}
 }
